refactor: rename game_chunks -> game_terrain
