+++ src/core/modules/game_terrain.js
